refactor: split index.js into modular components
- Extract JXAExecutor utility class for script execution
- Extract BuildLogParser utility for log parsing and derived data handling
- Extract PathValidator utility for project/file path validation
- Extract BuildTools class for build, clean, test, run, debug operations
- Extract ProjectTools class for project operations (open, schemes, destinations)
- Extract InfoTools class for workspace and project info operations
- Create XcodeServer main class with tool registration and MCP handlers
- Update index.js to be a simple entry point extending XcodeServer
- Maintain backward compatibility with all existing method signatures
- All critical tests passing (parameter validation, functionality)

This refactoring improves code organization, maintainability, and testability
by separating concerns into focused, single-responsibility modules.
